# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "client_message_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^[A-Za-z0-9._:-]{1,64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Retries with the same id reuse the existing job instead of starting another.",
            "title": "Client Message Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "dispatch": {
            "default": "inline",
            "description": "inline runs the job in this request. worker leaves it queued for the worker.",
            "enum": [
              "inline",
              "worker"
            ],
            "title": "Dispatch",
            "type": "string"
          },
          "horizon": {
            "default": "2-5 sessions",
            "maxLength": 64,
            "minLength": 1,
            "title": "Horizon",
            "type": "string"
          },
          "message": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "recording_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recording Id"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "default": "5m",
            "enum": [
              "1m",
              "5m",
              "15m",
              "1h",
              "4h",
              "1d"
            ],
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "message",
          "symbol"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Research-only API for the Trading Research Workspace. Serves shared contracts, fixture market data and (from Stage 1B) chat, runs and artifacts. There is no broker order-write capability.",
    "title": "Trading Research Workspace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chat": {
      "post": {
        "operationId": "postChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Chat",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}